requirements.txt
# ──────────────────────────────────────────────────────────────
#  Mini Data Platform — Python Dependencies
# ──────────────────────────────────────────────────────────────

# Data generation
faker==22.2.0

# Data processing
pandas==2.2.0

# MinIO client
minio==7.2.3

# PostgreSQL
psycopg2-binary==2.9.9

# Airflow providers
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-amazon==8.18.0
